# public/openapi.yaml
openapi: 3.1.0
info:
  title: CRM Follow-up Agent API
  description: API for ChatGPT Custom Actions to manage leads follow-up, tasks, and notes securely.
  version: 1.0.0
servers:
  - url: https://{YOUR_VERCEL_DOMAIN}
    description: Production Server
security:
  - bearerAuth: []
paths:
...
  /api/gpt/calendar/events:
    get:
      summary: List calendar events
      description: Returns a list of scheduled events from the user's calendar.
      operationId: getCalendarEvents
      parameters:
        - name: start_date
          in: query
          description: Filter events starting from this date (ISO 8601).
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          description: Filter events ending before this date (ISO 8601).
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Successful response
    post:
      summary: Schedule a new calendar event
      description: Creates a new event (meeting, call, visit) in the calendar.
      operationId: createCalendarEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - start_time
                - end_time
              properties:
                title:
                  type: string
                description:
                  type: string
                start_time:
                  type: string
                  format: date-time
                end_time:
                  type: string
                  format: date-time
                event_type:
                  type: string
                  enum: [meeting, call, visit, viewing, task, deadline, other]
                lead_id:
                  type: string
                  format: uuid
                location:
                  type: string
      responses:
        '201':
          description: Event scheduled successfully
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key